Move task processor database settings validation to the right place
